+++ src/features/admin/api.ts
@@ -30,7 +30,6 @@ import {
   type ListingInput,
   listingsTable,
 } from "#shared/db/listings.ts";
-import { validateListingFieldValues } from "#shared/listing-field-validators.ts";
 import {
   deleteOrphanedAddOnError,
   generateUniqueListingSlug,
@@ -519,16 +518,6 @@ const hydrateListingGroupIds = async (
   );
 };
 
-/** The JSON API builds a `ListingInput` directly, so — like the catalog import
- * — it must run the form's per-field value rules (which `validateListingInput`
- * itself does not) in addition to the cross-entity checks. */
-const validateListingApiInput = async (
-  input: ListingInput,
-  existingId?: number,
-): Promise<string | null> =>
-  validateListingFieldValues(input) ??
-  (await validateListingInput(input, existingId));
-
 const listingApiRoutes = defineCrudApi<
   Listing,
   ListingInput,
@@ -577,7 +566,7 @@ const listingApiRoutes = defineCrudApi<
   table: listingsTable,
   toCreateInput: bodyToCreateInput,
   toUpdateInput: bodyToUpdateInput,
-  validate: validateListingApiInput,
+  validate: validateListingInput,
 });
 
 export const adminApiRoutes = {
