+++ src/features/admin/catalog-transfer/schema.ts
@@ -34,8 +34,11 @@ import {
  */
 const isStorableDatetime = (value: string): boolean => {
   if (value === "") return true;
+  // Anchored end ($) so trailing junk ("…T00:00not-a-zone") is rejected rather
+  // than silently emptied by the storage normaliser: optional seconds, optional
+  // fractional seconds, and an optional Z / ±HH:MM offset are the only tails.
   const m = value.match(
-    /^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2})(?::(\d{2}))?/,
+    /^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2})(?::(\d{2}))?(?:\.\d+)?(?:Z|[+-]\d{2}:\d{2})?$/,
   );
   if (!m) return false;
   const y = Number(m[1]);
