+++ src/ui/templates/fields.ts
@@ -5,6 +5,7 @@
 import * as v from "valibot";
 import { t } from "#i18n";
 import { formatCurrency, getDecimalPlaces } from "#shared/currency.ts";
+import { DAY_NAMES } from "#shared/dates.ts";
 import { isUpdateTier } from "#shared/db/built-sites.ts";
 import { CONFIG_KEYS, settings } from "#shared/db/settings.ts";
 import type { FormParams } from "#shared/form-data.ts";
@@ -15,14 +16,6 @@ import {
   MAX_IMAGE_SIZE,
   MAX_TEXTAREA_LENGTH,
 } from "#shared/limits.ts";
-import {
-  splitCsv,
-  VALID_DAY_NAMES,
-  validateBookableDays,
-  validateDurationDays,
-  validateListingFields,
-  validateListingType,
-} from "#shared/listing-field-validators.ts";
 import {
   mergeListingFields,
   parseListingFields,
@@ -38,6 +31,8 @@ import {
   type AdminLevel,
   type ContactField,
   type ContactInfo,
+  isContactField,
+  isListingType,
   type ListingFields,
   type ListingType,
   MAX_DURATION_DAYS,
@@ -251,20 +246,60 @@ const getUsernameFieldBase = (): Field => ({
   type: "text",
 });
 
-/** Re-exported so existing importers keep resolving these listing-field
- * validators through the field module. */
-export {
-  splitCsv,
-  VALID_DAY_NAMES,
-  validateBookableDays,
-  validateDurationDays,
-  validateListingFields,
-  validateListingType,
+/** Validate listing fields setting (comma-separated contact field names) */
+const validateListingFields = (value: string): string | null => {
+  const parts = value
+    .split(",")
+    .map((v) => v.trim())
+    .filter((v) => v);
+  for (const part of parts) {
+    if (!isContactField(part)) {
+      return t("fields.validation.invalid_contact_field", { part });
+    }
+  }
+  return null;
+};
+
+/** Validate listing type setting */
+const validateListingType = (value: string): string | null => {
+  if (!isListingType(value)) {
+    return t("fields.validation.listing_type");
+  }
+  return null;
 };
 
 /** Validate a built site's update channel (alpha/beta/release) */
 const validateUpdateTier = (value: string): string | null =>
   isUpdateTier(value) ? null : t("fields.validation.update_tier");
+
+/** Valid day names for bookable_days (Monday-first for display) */
+export const VALID_DAY_NAMES = [...DAY_NAMES.slice(1), DAY_NAMES[0]!];
+
+/** Check if a string is a valid day name */
+const isValidDayName = (s: string): boolean =>
+  (VALID_DAY_NAMES as readonly string[]).includes(s);
+
+/** Split a comma-separated string into trimmed, non-empty tokens */
+export const splitCsv = (value: string): string[] =>
+  value
+    .split(",")
+    .map((d) => d.trim())
+    .filter((d) => d);
+
+/** Validate bookable days (comma-separated day names) */
+export const validateBookableDays = (value: string): string | null => {
+  const days = splitCsv(value);
+  if (days.length === 0) return t("fields.validation.days_required");
+  for (const day of days) {
+    if (!isValidDayName(day)) {
+      return t("fields.validation.invalid_day", {
+        day,
+        valid: VALID_DAY_NAMES.join(", "),
+      });
+    }
+  }
+  return null;
+};
 
 /** Shared formatting hint linking to the admin guide */
 // Links to the standalone formatting-help page (not the staff-only full guide),
@@ -293,7 +328,7 @@ const buildDescriptionField = (hint: string, hintHtml?: string): Field => ({
 });
 
 /** Validate a datetime value is parseable */
-export const validateDatetime = (value: string): string | null =>
+const validateDatetime = (value: string): string | null =>
   isValidDatetime(value) ? null : t("fields.validation.datetime");
 
 /** Build a "hidden" visibility checkbox field for a listing or group. */
@@ -411,7 +446,19 @@ export const getListingFields = (): Field[] => [
     min: 1,
     name: "duration_days",
     type: "number",
-    validate: validateDurationDays,
+    validate: (value: string): string | null => {
+      // validateSingleField only calls this when the value is non-empty, so
+      // the empty-string case never reaches here.
+      const parsed = Number(value);
+      if (!Number.isInteger(parsed)) {
+        return t("fields.validation.duration_whole");
+      }
+      if (parsed < 1) return t("fields.validation.duration_min");
+      if (parsed > MAX_DURATION_DAYS) {
+        return t("fields.validation.duration_max", { max: MAX_DURATION_DAYS });
+      }
+      return null;
+    },
   },
   {
     hint: t("fields.listing.customisable_days_hint"),
